refactor : line to stream
