+++ src/test/java/nextstep/ladder/domain/LadderTest.java
@@ -1,4 +1,4 @@
-package nextstep.ladder;
+package nextstep.ladder.domain;
 
 import nextstep.ladder.domain.Ladder;
 import org.junit.jupiter.api.Test;
